Update remainder to round to 2 decimal spaces

+++ src/components/addCharge/addCharge.js
@@ -34,7 +34,7 @@ class AddCharge extends Component {
   }
 
   updateAmount(amount) {
-    this.setState({ amount: parseInt(amount) });
+    this.setState({ amount: parseFloat(amount) });
   }
 
   updateCategory(category) {
